// src/core/state/ducks/track/track.types.ts
enum ActionTypes {
  SET_TRACK_ID = "track/SET_TRACK_ID",
  FETCH_META_PENDING = "track/FETCH_META_PENDING",
  FETCH_META_SUCCESS = "track/FETCH_META_SUCCESS",
  FETCH_META_ERROR = "track/FETCH_META_ERROR",
  SET_AUDIO_SOURCE = "track/SET_AUDIO_SOURCE",
  SET_AUDIO_DURATION = "track/SET_AUDIO_DURATION",
  SET_AUDIO_STATUS = "track/SET_AUDIO_STATUS",
  TOOGLE_AUDIO_STATUS = "track/TOOGLE_AUDIO_STATUS",
  SET_AUDIO_CURRENT_TIME = "track/SET_AUDIO_CURRENT_TIME",
  TRANSFER_META_SUCCESS = "track/TRANSFER_META_SUCCESS"
}

export default ActionTypes;